player_sessions, make use of argon2id instead of argon2i securerandomblake2b with monocypher for session random number

+++ common/NetEngine/CMessage.cpp
@@ -91,7 +91,7 @@ namespace NetEngine
     }
     uint32_t CMessage::GetDataSize()
     {
-        return GetFullSize() - minSize();
+        return GetFullSize() - static_cast<uint32_t>(minSize());
     }
     uint32_t CMessage::GetFullSize()
     {
